[*] Changed ret code from unsigned int to signed it for BLISS algo

+++ crypto/dap_enc_bliss.h
@@ -16,10 +16,10 @@ void dap_enc_sig_bliss_key_new_generate(struct dap_enc_key * key, const void *ke
                                     size_t kex_size, const void * seed, size_t seed_size,
                                     size_t key_size);
 
-size_t dap_enc_sig_bliss_get_sign(struct dap_enc_key * key,const void * msg,
+int dap_enc_sig_bliss_get_sign(struct dap_enc_key * key,const void * msg,
                                   const size_t msg_size, void * signature, const size_t signature_size);
 
-size_t dap_enc_sig_bliss_verify_sign(struct dap_enc_key * key,const void * msg,
+int dap_enc_sig_bliss_verify_sign(struct dap_enc_key * key,const void * msg,
                                      const size_t msg_size, void * signature, const size_t signature_size);
 
 void dap_enc_sig_bliss_key_delete(struct dap_enc_key *key);
